requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
python-dotenv
pydantic-settings
paramiko
passlib[bcrypt]
python-jose[cryptography]
celery
email-validator
itsdangerous
pytest
pytest-asyncio
httpx
aiosqlite